LPS-91769 Inline DeploymentStatement
